Rename wayland_buffer.{c,h} to buffer.{c,h}
